decouple business logic from UI
