+++ src/daxxl/gui/gui.py
@@ -1,32 +1,23 @@
 import asyncio
 import sys
-from datetime import datetime, timezone
-from enum import Enum
 from pathlib import Path
 from tkinter import DISABLED, NORMAL, PhotoImage, Tk, Widget
 from tkinter.messagebox import showerror, showinfo, showwarning
-from typing import Any, Awaitable, Callable, Dict, List, Optional, Set, Tuple, Union
+from typing import Any, Callable, Dict, List, Union
 
-import httpx
-from colorama import Fore
 from ttkthemes import ThemedTk
 
-from daxxl.assembler.assembler import ReleaseAssembler
-from daxxl.defs import Archive, ModSource, Position, Side
-from daxxl.exceptions import NoModAssetFound, ReleaseNotFoundException
+from daxxl.defs import Archive, Position, Side
+from daxxl.exceptions import ReleaseNotFoundException, SideAlreadySetException
 from daxxl.gtnh_logger import get_logger
 from daxxl.gui.exclusion.exclusion_panel import ExclusionPanel, ExclusionPanelCallback
 from daxxl.gui.external.external_panel import ExternalPanel, ExternalPanelCallback
 from daxxl.gui.github.github_panel import GithubPanel, GithubPanelCallback
 from daxxl.gui.lib.decorators import with_error_dialog
-from daxxl.gui.lib.progress_bar import CustomProgressBar
 from daxxl.gui.modpack.modpack_panel import ModpackPanel, ModpackPanelCallback
-from daxxl.models.gtnh_config import GTNHConfig
 from daxxl.models.gtnh_release import GTNHRelease
-from daxxl.models.gtnh_version import GTNHVersion
 from daxxl.models.mod_info import GTNHModInfo
-from daxxl.models.mod_version_info import ModVersionInfo
-from daxxl.modpack_manager import GTNHModpackManager
+from daxxl.release_controller import DevRelease, ReleaseController
 
 logger = get_logger(__name__)
 
@@ -65,13 +56,11 @@ class App:
         """
         await self.instance.run()
 
-class DevRelease(str, Enum):
-    DAILY = "daily"
-    EXPERIMENTAL = "experimental"
 
 class Window(ThemedTk, Tk):
     """
-    Main class for the GUI.
+    Main class for the GUI. Only handles widgets, dialogs and the event loop;
+    all the business logic is delegated to the ReleaseController.
     """
 
     def __init__(self, themed: bool = False) -> None:
@@ -86,8 +75,6 @@ class Window(ThemedTk, Tk):
             ThemedTk.__init__(self, theme=theme)
         else:
             Tk.__init__(self)
-        self._client: Optional[httpx.AsyncClient] = None
-        self._modpack_manager: Optional[GTNHModpackManager] = None
         self._run: bool = True
         self.title("DreamAssemblerXXL")
         self.iconphoto(False, PhotoImage(file=ICON))
@@ -96,22 +83,8 @@ class Window(ThemedTk, Tk):
         self.ypadding: int = 0
         self.minsize(1000, 600)  # Minimum to even see all the buttons
 
-        self.github_mods: Dict[
-            str, ModVersionInfo
-        ] = {}  # name <-> version of github mods mappings for the current release
-        self.gtnh_config: str = ""  # modpack asset version
-        self.external_mods: Dict[
-            str, ModVersionInfo
-        ] = {}  # name <-> version of external mods mappings for the current release
-        self.version: str = ""  # modpack release name
-        self.last_version: Optional[str] = None  # last version of the release
-
-        self.download_error_list: List[str] = []  # list of errors happened during the download of a release's assets
-
         self.init: bool = False
         self.protocol("WM_DELETE_WINDOW", lambda: asyncio.ensure_future(self.close_app()))
-
-        self.delta_progress: float = 0  # progression between 2 tasks (in %) for the global progress bar
 
         # frame for the modpack handling
         modpack_panel_callbacks: ModpackPanelCallback = ModpackPanelCallback(
@@ -129,7 +102,7 @@ class Window(ThemedTk, Tk):
             update_all=lambda: asyncio.ensure_future(self.assemble_all()),
             update_beta=lambda: asyncio.ensure_future(self.assemble_beta()),
             generate_changelog=lambda: asyncio.ensure_future(self.generate_changelog()),
-            generate_cf_files=lambda: asyncio.ensure_future(self.generetate_intermediate_cf_files()),
+            generate_cf_files=lambda: asyncio.ensure_future(self.generate_intermediate_cf_files()),
             load=lambda release_name: asyncio.ensure_future(self.load_gtnh_version(release_name)),
             delete=lambda release_name: asyncio.ensure_future(self.delete_gtnh_version(release_name)),
             add=lambda release_name, previous_version: asyncio.ensure_future(
@@ -152,22 +125,29 @@ class Window(ThemedTk, Tk):
             self.modpack_list_frame.action_frame.progress_bar_current_task.reset
         )
 
+        self.controller: ReleaseController = ReleaseController(
+            progress_callback=self.progress_callback,
+            global_callback=self.global_callback,
+            global_reset_callback=self.global_reset_callback,
+            current_task_reset_callback=self.current_task_reset_callback,
+        )
+
         # frame for the github mods
         github_panel_callbacks: GithubPanelCallback = GithubPanelCallback(
-            get_gtnh_callback=self._get_modpack_manager,
-            get_github_mods_callback=self.get_github_mods,
-            set_mod_version=self.set_github_mod_version,
+            get_gtnh_callback=self.controller.get_modpack_manager,
+            get_github_mods_callback=self.controller.get_github_mods,
+            set_mod_version=self.controller.set_github_mod_version,
             set_mod_side=lambda name, side: asyncio.ensure_future(self.set_github_mod_side(name, side)),
             set_mod_side_default=lambda name, side: asyncio.ensure_future(
                 self.set_mod_side_default(name, side)
             ),
-            set_modpack_version=self.set_modpack_version,
+            set_modpack_version=self.controller.set_modpack_version,
             update_current_task_progress_bar=self.progress_callback,
             update_global_progress_bar=self.global_callback,
             reset_current_task_progress_bar=self.current_task_reset_callback,
             reset_global_progress_bar=self.global_reset_callback,
-            add_mod_in_memory=self._add_mod,
-            del_mod_in_memory=self._del_github_mod,
+            add_mod_in_memory=self.controller.add_github_mod,
+            del_mod_in_memory=self.controller.del_github_mod,
         )
 
         self.github_panel: GithubPanel = GithubPanel(
@@ -177,16 +157,16 @@ class Window(ThemedTk, Tk):
         # frame for the external mods
 
         external_panel_callbacks: ExternalPanelCallback = ExternalPanelCallback(
-            set_mod_version=self.set_external_mod_version,
+            set_mod_version=self.controller.set_external_mod_version,
             set_mod_side=lambda name, side: asyncio.ensure_future(self.set_external_mod_side(name, side)),
             set_mod_side_default=lambda name, side: asyncio.ensure_future(
                 self.set_mod_side_default(name, side)
             ),
-            get_gtnh_callback=self._get_modpack_manager,
-            get_external_mods_callback=self.get_external_mods,
+            get_gtnh_callback=self.controller.get_modpack_manager,
+            get_external_mods_callback=self.controller.get_external_mods,
             toggle_freeze=self.trigger_toggle,
-            add_mod_in_memory=self._add_external_mod,
-            del_mod_in_memory=self._del_external_mod,
+            add_mod_in_memory=self.controller.add_external_mod,
+            del_mod_in_memory=self.controller.del_external_mod,
             refresh_external_modlist=self.refresh_external_mods,
         )
 
@@ -195,8 +175,8 @@ class Window(ThemedTk, Tk):
         )
 
         exclusion_client_callbacks: ExclusionPanelCallback = ExclusionPanelCallback(
-            add=lambda exclusion: asyncio.ensure_future(self.add_exclusion("client", exclusion)),
-            delete=lambda exclusion: asyncio.ensure_future(self.del_exclusion("client", exclusion)),
+            add=lambda exclusion: asyncio.ensure_future(self.controller.add_exclusion("client", exclusion)),
+            delete=lambda exclusion: asyncio.ensure_future(self.controller.del_exclusion("client", exclusion)),
         )
 
         # frame for the client file exclusions
@@ -205,8 +185,8 @@ class Window(ThemedTk, Tk):
         )
 
         exclusion_server_callbacks: ExclusionPanelCallback = ExclusionPanelCallback(
-            add=lambda exclusion: asyncio.ensure_future(self.add_exclusion("server", exclusion)),
-            delete=lambda exclusion: asyncio.ensure_future(self.del_exclusion("server", exclusion)),
+            add=lambda exclusion: asyncio.ensure_future(self.controller.add_exclusion("server", exclusion)),
+            delete=lambda exclusion: asyncio.ensure_future(self.controller.del_exclusion("server", exclusion)),
         )
 
         # frame for the server side exclusions
@@ -219,44 +199,6 @@ class Window(ThemedTk, Tk):
 
         self.toggled: bool = True  # state variable indicating if the widgets are disabled or not
 
-    def _add_mod(self, name: str, version: str) -> None:
-        """
-        add a mod to inmemory github modlist.
-
-        :param name: mod name
-        :param version: mod version
-        :return: None
-        """
-        self.github_mods[name] = ModVersionInfo(version=version)
-
-    def _del_github_mod(self, name: str) -> None:
-        """
-        remove a mod from inmemory github modlist.
-
-        :param name: mod name
-        :return: None
-        """
-        del self.github_mods[name]
-
-    def _add_external_mod(self, name: str, version: str) -> None:
-        """
-        add a mod to inmemory external modlist.
-
-        :param name: mod name
-        :param version: mod version
-        :return: None
-        """
-        self.external_mods[name] = ModVersionInfo(version=version)
-
-    def _del_external_mod(self, name: str) -> None:
-        """
-        remove a mod from inmemory external modlist.
-
-        :param name: mod name
-        :return: None
-        """
-        del self.external_mods[name]
-
     def trigger_toggle(self) -> None:
         """
         Enable/disable the widgets that can be toggled.
@@ -264,7 +206,6 @@ class Window(ThemedTk, Tk):
         :return: None
         """
         self.toggled = not self.toggled
-        # print(f"{'toggled' if not self.toggled else 'untoggled'}") # debug
         self.toggle(self)
 
     def toggle(self, widget: Any) -> None:
@@ -291,51 +232,35 @@ class Window(ThemedTk, Tk):
         title="An error occured during the generation of the changelog",
         message=lambda self: (
                 f"An error occured during the generation of the changelog from "
-                f"{self.last_version} to {self.version}.\nPlease check the logs for more information."
+                f"{self.controller.last_version} to {self.controller.version}."
+                "\nPlease check the logs for more information."
         ),
     )
     async def generate_changelog(self) -> None:
         """
-        Method used to trigger the assembling of the client archive corresponding to the provided source.
+        Callback used to generate the changelog of the loaded release.
 
         :return: None
         """
-        global_callback: Callable[[float, str], None] = (
-            self.modpack_list_frame.action_frame.progress_bar_global.add_progress
-        )
-
-        self.set_progress(100 / 2)
         self.trigger_toggle()
-        release_assembler: ReleaseAssembler = await self.pre_assembling()
-        release_assembler.generate_changelog()
-        global_callback(self.get_progress(), f"Generate changelog from {self.last_version} to {self.version}")
+        await self.controller.generate_changelog()
         self.trigger_toggle()
-
 
     @with_error_dialog(
         title="An error occured during the generation of the intermediate curseforge files",
         message="An error occured during the generation of the intermediate curseforge files."
                 "\nPlease check the logs for more information."
     )
-    async def generetate_intermediate_cf_files(self) -> None:
+    async def generate_intermediate_cf_files(self) -> None:
         """
-        Method used to generate curseforge intermediate files.
+        Callback used to generate curseforge intermediate files.
 
         :return: None
         """
-        global_callback: Callable[[float, str], None] = (
-            self.modpack_list_frame.action_frame.progress_bar_global.add_progress
-        )
-        task_callback_object: CustomProgressBar = self.modpack_list_frame.action_frame.progress_bar_current_task
-
-        self.set_progress(100 / 3)
         self.trigger_toggle()
-
-        release_assembler = await self.pre_assembling()
-        global_callback(self.get_progress(), "Generating the dependencies.json")
-        await release_assembler.curse_assembler.generate_json_dep(task_callback_object)
-        global_callback(self.get_progress(), "Generating the archive containing the mods to upload")
-        release_assembler.curse_assembler.generate_mods_to_upload(task_callback_object)
+        await self.controller.generate_intermediate_cf_files(
+            self.modpack_list_frame.action_frame.progress_bar_current_task
+        )
         self.trigger_toggle()
 
     @with_error_dialog(
@@ -348,82 +273,13 @@ class Window(ThemedTk, Tk):
     )
     async def assemble_release(self, side: Side, archive_type: Archive) -> None:
         """
-        Method used to trigger the assembling of the client archive corresponding to the provided source.
+        Callback used to trigger the assembling of the archive corresponding to the provided side and type.
 
         :return: None
         """
-        global_callback: Callable[[float, str], None] = (
-            self.modpack_list_frame.action_frame.progress_bar_global.add_progress
-        )
-
-        self.set_progress(100 / 2)
         self.trigger_toggle()
-        release_assembler: ReleaseAssembler = await self.pre_assembling()
-        assembler_dict: Dict[Archive, Callable[[Side, bool], Awaitable[None]]] = {
-            Archive.ZIP: release_assembler.assemble_zip,
-            Archive.MMC: release_assembler.assemble_mmc,
-            Archive.MODRINTH: release_assembler.assemble_modrinth,
-            Archive.CURSEFORGE: release_assembler.assemble_curse,
-            Archive.TECHNIC: release_assembler.assemble_technic,
-        }
-        global_callback(self.get_progress(), f"Assembling {side.value} {archive_type.value} archive")
-        await assembler_dict[archive_type](side=side, verbose=True)  # type: ignore
+        await self.controller.assemble_release(side, archive_type)
         self.trigger_toggle()
-
-
-    async def pre_assembling(self) -> ReleaseAssembler:
-        """
-        Method to downloads the mods before constructing the ReleaseAssembler object.
-
-        :return: the ReleaseAssembler object constructed
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        release: GTNHRelease = GTNHRelease(
-            version=self.version,
-            config=self.gtnh_config,
-            github_mods=self.github_mods,
-            external_mods=self.external_mods,
-            last_version=self.last_version,
-        )
-        logger.info(f"version: {self.version}")
-
-        # clean the previous state of the progress bars
-        self.global_reset_callback()
-        self.current_task_reset_callback()
-
-        self.global_callback(self.get_progress(), "Downloading assets")
-        await gtnh.download_release(
-            release, download_callback=self.progress_callback, error_callback=self.add_error_message
-        )
-        self.current_task_reset_callback()
-
-        if len(self.download_error_list) > 0:
-            error = "The following error(s) happened during the downloading of assets:\n" + "\n".join(
-                self.download_error_list
-            )
-
-            showerror("Error(s) happened during the downloading of assets", error)
-            self.download_error_list = []
-            self.trigger_toggle()
-
-            raise NoModAssetFound(error)
-
-        return ReleaseAssembler(
-            gtnh,
-            release,
-            task_callback=self.progress_callback,
-            global_callback=self.global_callback,
-            current_task_reset_callback=self.current_task_reset_callback,
-        )
-
-    def add_error_message(self, error_message: str) -> None:
-        """
-        Method used as error callback when an error happens during the download of a mod/release.
-
-        :param error_message: the error message to add to the error list
-        :return: None
-        """
-        self.download_error_list.append(error_message)
 
     @with_error_dialog(
         title="An error occured during the update of the assembling of the archives",
@@ -432,33 +288,12 @@ class Window(ThemedTk, Tk):
     )
     async def assemble_all(self) -> None:
         """
-        Assemble all the archives for a full update.
+        Callback used to assemble all the archives for a full update.
 
         :return: None
         """
-        global_callback: Callable[[float, str], None] = (
-            self.modpack_list_frame.action_frame.progress_bar_global.add_progress
-        )
-
         self.trigger_toggle()
-
-        self.set_progress(100 / (1 + 5 + 1))  # download + archives for client + archive for server
-        release_assembler: ReleaseAssembler = await self.pre_assembling()
-
-        release_assembler.set_progress(self.get_progress())
-
-        await release_assembler.assemble(Side.CLIENT, verbose=True)
-
-        await self.assemble_release(Side.CLIENT_JAVA9, Archive.MMC)
-        await self.assemble_release(Side.SERVER_JAVA9, Archive.ZIP)
-
-        # todo: redo the bar resets less hacky: they are spread all over the place and it's inconsistent
-        if release_assembler.current_task_reset_callback is not None:
-            release_assembler.current_task_reset_callback()
-
-        global_callback(self.get_progress(), f"Assembling {Side.SERVER} {Archive.ZIP} archive")
-        await release_assembler.assemble_zip(Side.SERVER, verbose=True)
-
+        await self.controller.assemble_all()
         self.trigger_toggle()
 
     @with_error_dialog(
@@ -468,239 +303,61 @@ class Window(ThemedTk, Tk):
     )
     async def assemble_beta(self) -> None:
         """
-        Assemble all the archives for a beta/RC update.
+        Callback used to assemble all the archives for a beta/RC update.
 
         :return: None
         """
         self.trigger_toggle()
-
-        self.set_progress(100 / (1 + 3 + 2))  # download + archives for client + archive for server
-        release_assembler: ReleaseAssembler = await self.pre_assembling()
-
-        release_assembler.set_progress(self.get_progress())
-
-        # zip archives
-        await self.assemble_release(Side.CLIENT, Archive.ZIP)
-        await self.assemble_release(Side.SERVER, Archive.ZIP)
-        await self.assemble_release(Side.SERVER_JAVA9, Archive.ZIP)
-
-        # MMC archives
-        await self.assemble_release(Side.CLIENT, Archive.MMC)
-        await self.assemble_release(Side.CLIENT_JAVA9, Archive.MMC)
-
+        await self.controller.assemble_beta()
         self.trigger_toggle()
 
-    def set_progress(self, delta_progress: float) -> None:
-        """
-        Setter for self.delta_progress.
-
-        :param delta_progress: new progress
-        :return: None
-        """
-        self.delta_progress = delta_progress
-
-    def get_progress(self) -> float:
-        """
-        Setter for self.delta_progress.
-
-        :return: the current delta progress
-        """
-        return self.delta_progress
-
-    async def add_exclusion(self, side: str, exclusion: str) -> None:
-        """
-        Method used to add an file exclusion to the corresponding side's exclusion list.
-
-        :param side: side of the modpack
-        :param exclusion: the string corresponding to the file exclusion
-        :return: None
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        gtnh.add_exclusion(side, exclusion)
-        gtnh.save_modpack()
-
-    async def del_exclusion(self, side: str, exclusion: str) -> None:
-        """
-        Method used to add an file exclusion to the corresponding side's exclusion list.
-
-        :param side: side of the modpack
-        :param exclusion: the string corresponding to the file exclusion
-        :return: None
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        gtnh.delete_exclusion(side, exclusion)
-        gtnh.save_modpack()
-
-    async def get_modpack_exclusions(self, side: str) -> List[str]:
-        """
-        Method used to gather the file exclusion list of the modpack corresponding to the provided side.
-
-        :param side: side of the pack
-        :return: list of strings corresponding to the file exclusions
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        if side == "client":
-            return sorted([exclusion for exclusion in gtnh.mod_pack.client_exclusions])
-        elif side == "server":
-            return sorted([exclusion for exclusion in gtnh.mod_pack.server_exclusions])
-        else:
-            raise ValueError(f"side {side} is an invalid side")
-
-    async def _set_mod_side(
-            self,
-            mods: Dict[str, ModVersionInfo],
-            mod_name: str,
-            side: Side,
-            get_default_version: Callable[[], str],
-    ) -> None:
-        """
-        Change the side of a mod in the given dict (github_mods or external_mods),
-        creating the entry if it does not exist or deleting the entry if side is Side.NONE.
-
-        :param mods: the given dict (github_mods or external_mods)
-        :param mod_name: mod name
-        :param side: new Side
-        :param get_default_version: callback giving the default value if the entry does not exist in the dict
-        :return: None
-        """
-        previous_side = mods[mod_name].side if mod_name in mods else Side.NONE
-        if previous_side == side:
-            showwarning("Side already set up", f"{mod_name}'s side is already set to {side}")
-            return
-
-        if side == Side.NONE:
-            del mods[mod_name]
-        elif previous_side == Side.NONE:
-            mods[mod_name] = ModVersionInfo(version=get_default_version(), side=side)
-        else:
-            mods[mod_name].side = side
     async def set_github_mod_side(self, mod_name: str, side: Side) -> None:
         """
-        Method used to set the side of a github mod.
+        Callback used to set the side of a github mod.
 
         :param mod_name: the mod name
         :param side: side of the pack, None if use default
         :return: None
         """
-        await self._set_mod_side(
-            self.github_mods, mod_name, side, lambda: self.github_panel.mod_info_frame.version.get()
-        )
-
-    async def set_mod_side_default(self, mod_name: str, side: str) -> None:
-        """
-        Set the mod side to the given side no matter what is its source (github or external).
-
-        :param mod_name: mod name
-        :param side: the default side to apply
-        :return: None
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        previous_side: Side = gtnh.assets.get_mod(mod_name).side
-        if previous_side == side:
-            showwarning(
-                "Side already set up",
-                f"{mod_name}'s side is already set on {side}",
-            )
-            return
-
-        if not gtnh.set_mod_side(mod_name, side):
-            showerror(
-                "Error setting up the side of the mod",
-                f"Error during the process of setting up {mod_name}'s side to {side}. Check the logs for more details",
-            )
-            return
+        try:
+            self.controller.set_github_mod_side(mod_name, side, self.github_panel.mod_info_frame.version.get)
+        except SideAlreadySetException as e:
+            showwarning("Side already set up", str(e))
 
     async def set_external_mod_side(self, mod_name: str, side: Side) -> None:
         """
-        Method used to set the side of an external mod.
+        Callback used to set the side of an external mod.
 
         :param mod_name: the mod name
         :param side: side of the pack
         :return: None
         """
-        await self._set_mod_side(
-            self.external_mods, mod_name, side, lambda: self.external_mod_frame.mod_info_frame.version.get()
-        )
+        try:
+            self.controller.set_external_mod_side(mod_name, side, self.external_mod_frame.mod_info_frame.version.get)
+        except SideAlreadySetException as e:
+            showwarning("Side already set up", str(e))
 
-    def set_github_mod_version(self, github_mod_name: str, mod_version: str) -> None:
+    async def set_mod_side_default(self, mod_name: str, side: str) -> None:
         """
-        Callback used when a github mod version is selected.
+        Callback used to set the default side of a mod no matter what is its source (github or external).
 
-        :param github_mod_name: mod name
-        :param mod_version: mod version
+        :param mod_name: mod name
+        :param side: the default side to apply
         :return: None
         """
-        if github_mod_name in self.github_mods:
-            self.github_mods[github_mod_name].version = mod_version
-
-    def set_external_mod_version(self, external_mod_name: str, mod_version: str) -> None:
-        """
-        Callback used when an external mod version is selected.
-
-        :param external_mod_name: mod name
-        :param mod_version: mod version
-        :return: None
-        """
-        if external_mod_name in self.external_mods:
-            self.external_mods[external_mod_name].version = mod_version
-
-    def set_modpack_version(self, modpack_version: str) -> None:
-        """
-        Callback used when a modpack version is selected.
-
-        :param modpack_version: modpack version
-        :return: None
-        """
-        self.gtnh_config = modpack_version
-
-    async def get_repos(self) -> List[str]:
-        """
-        Method to grab all the repo names known.
-
-        :return: a list of github mod names
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        return [x.name for x in gtnh.assets.mods if x.source == ModSource.github]
-
-    def get_github_mods(self) -> Dict[str, ModVersionInfo]:
-        """
-        Getter for self.github_mods.
-
-        :return: self.github_mods
-        """
-        return self.github_mods
-
-    def get_external_mods(self) -> Dict[str, ModVersionInfo]:
-        """
-        Getter for self.external_mods.
-
-        :return: self.external_mods
-        """
-        return self.external_mods
-
-    async def _get_client(self) -> httpx.AsyncClient:
-        """
-        Internal getter for the httpx client instance, creating it if it doesn't exist.
-
-        :return: the httpx client instance
-        """
-        if self._client is None:
-            self._client = httpx.AsyncClient(http2=True)
-        return self._client
-
-    async def _get_modpack_manager(self) -> GTNHModpackManager:
-        """
-        Internal getter for the modpack manager instance, creating it if it doesn't exist.
-
-        :return: the modpack manager instance
-        """
-        if self._modpack_manager is None:
-            self._modpack_manager = GTNHModpackManager(await self._get_client())
-        return self._modpack_manager
+        try:
+            if not await self.controller.set_mod_side_default(mod_name, side):
+                showerror(
+                    "Error setting up the side of the mod",
+                    f"Error during the process of setting up {mod_name}'s side to {side}. "
+                    "Check the logs for more details",
+                )
+        except SideAlreadySetException as e:
+            showwarning("Side already set up", str(e))
 
     def _notify_errored_mods(
             self,
-            gtnh: GTNHModpackManager,
+            errored_mods: List[GTNHModInfo],
             title: str,
             success_message: str,
             warning_intro: str,
@@ -708,14 +365,12 @@ class Window(ThemedTk, Tk):
         """
         Show a warning if there was at least an errored mod, an ok message otherwise.
 
-        :param gtnh: the modpack manager instance
+        :param errored_mods: the mods needing attention
         :param title: dialogue title
         :param success_message: success message
         :param warning_intro: warning message prefix
         :return: None
         """
-        errored_mods = [mod for mod in gtnh.assets.mods if mod.needs_attention]
-
         if not errored_mods:
             showinfo(title, success_message)
             return
@@ -741,59 +396,30 @@ class Window(ThemedTk, Tk):
 
         :return: None
         """
-
-        self.global_reset_callback()
-        self.current_task_reset_callback()
-
         self.trigger_toggle()
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        global_delta_progress: float = 100 / (1 + 1)  # 1 for the syncing of the mods, 1 for update checks
-        await gtnh.update_all(
-            progress_callback=self.progress_callback,
-            global_progress_callback=lambda msg: self.global_callback(global_delta_progress, msg),
-        )
+        errored_mods: List[GTNHModInfo] = await self.controller.update_assets()
         self.trigger_toggle()
 
         self._notify_errored_mods(
-            gtnh,
+            errored_mods,
             title="assets updated successfully!",
             success_message="All the assets have been updated correctly!",
             warning_intro="The assets had been updated BUT:\n",
         )
 
-    async def _update_rolling_release(self, release_type: str) -> None:
+    async def _update_dev_release(self, release_type: str) -> None:
         """
-        update dev release (experimental/daily)
+        update dev release (experimental/daily).
 
         :param release_type: "experimental" or "daily"
         :return: None
         """
-        previous_release_name = f"previous_{release_type}"
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        existing_release = gtnh.get_release(release_type)
-        if not existing_release:
-            raise ReleaseNotFoundException(f"{release_type.capitalize()} release not found")
-
-        # 1 for the data download on github, 1 for the asset updates and 1 for the release update
-        global_delta_progress: float = 100 / (1 + 1 + 1)
-        release: GTNHRelease = await gtnh.update_release(
-            release_type,
-            existing_release=existing_release,
-            update_available=True,
-            progress_callback=self.progress_callback,
-            reset_progress_callback=self.current_task_reset_callback,
-            global_progress_callback=lambda msg: self.global_callback(global_delta_progress, msg),
-            last_version=previous_release_name,
-        )
-        gtnh.add_release(release, update=True)
-
-        # saving the previous snapshot
-        existing_release.version = previous_release_name
-        gtnh.add_release(existing_release, update=True)
-        gtnh.save_modpack()
+        self.trigger_toggle()
+        errored_mods: List[GTNHModInfo] = await self.controller.update_rolling_release(release_type)
+        self.trigger_toggle()
 
         self._notify_errored_mods(
-            gtnh,
+            errored_mods,
             title=f"updated the {release_type} release metadata",
             success_message=f"The {release_type} release metadata had been updated!",
             warning_intro=f"The {release_type} release metadata had been updated BUT:\n",
@@ -810,11 +436,7 @@ class Window(ThemedTk, Tk):
 
         :return: None
         """
-        self.current_task_reset_callback()
-        self.global_reset_callback()
-        self.trigger_toggle()
-        await self._update_rolling_release(DevRelease.EXPERIMENTAL.value)
-        self.trigger_toggle()
+        await self._update_dev_release(DevRelease.EXPERIMENTAL.value)
 
     @with_error_dialog(
         title="An error occured during the update of the daily build",
@@ -827,44 +449,7 @@ class Window(ThemedTk, Tk):
 
         :return: None
         """
-        self.current_task_reset_callback()
-        self.global_reset_callback()
-        self.trigger_toggle()
-        await self._update_rolling_release(DevRelease.DAILY.value)
-        self.trigger_toggle()
-
-    async def get_releases(self) -> List[GTNHRelease]:
-        """
-        Method used to return a list of known releases with valid metadata.
-        The list is sorted in ascending order (from oldest to the latest).
-
-        :return: a sorted list of all the gtnh releases availiable
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-
-        releases: List[GTNHRelease] = []
-
-        # if there is any release, chose last
-        if len(gtnh.mod_pack.releases) > 0:
-            # gtnh.mod_pack.releases is actually a set of the release names
-            for release_name in gtnh.mod_pack.releases:
-                release: Optional[GTNHRelease] = gtnh.get_release(release_name)
-
-                # discarding all the None releases, as it means the json data couldn't be loaded
-                if release is not None:
-                    releases.append(release)
-
-            # sorting releases by date. Some manifests store last_updated as offset-naive and others as
-            # offset-aware; normalize naive timestamps to UTC so they can be compared with each other.
-            def _sort_key(release_object: GTNHRelease) -> datetime:
-                last_updated = release_object.last_updated
-                if last_updated.tzinfo is None:
-                    return last_updated.replace(tzinfo=timezone.utc)
-                return last_updated
-
-            releases = sorted(releases, key=_sort_key)
-
-        return releases
+        await self._update_dev_release(DevRelease.DAILY.value)
 
     async def load_gtnh_version(self, release: Union[GTNHRelease, str], init: bool = False) -> None:
         """
@@ -874,23 +459,9 @@ class Window(ThemedTk, Tk):
         :param init: bool indicating if this is done manually or at init
         :return: None
         """
-        release_object: Optional[GTNHRelease]
-
-        if isinstance(release, str):
-            gtnh: GTNHModpackManager = await self._get_modpack_manager()
-            release_object = gtnh.get_release(release)
-        else:
-            release_object = release
-
-        if release_object is not None:
-            release_object = await self.strip_disabled_mods(release_object)
-            self.github_mods = release_object.github_mods
-            self.gtnh_config = release_object.config
-            self.external_mods = release_object.external_mods
-            self.version = release_object.version
-            self.last_version = release_object.last_version
-            logger.info(f"Loaded pack version {Fore.CYAN}{release_object.version}{Fore.RESET} in memory.")
-        else:
+        try:
+            release_object: GTNHRelease = await self.controller.load_gtnh_version(release)
+        except ReleaseNotFoundException:
             showerror("incorrect version detected", f"modpack version {release} doesn't exist")
             return
 
@@ -898,57 +469,7 @@ class Window(ThemedTk, Tk):
             showinfo("version loaded successfully!", f"modpack version {release_object.version} loaded successfully!")
         else:
             # display the loaded version at boot
-            self.modpack_list_frame.modpack_list.set_loaded_version(self.version)
-
-    async def strip_disabled_mods(self, release: GTNHRelease) -> GTNHRelease:
-        """
-        Method used to strip the disabled mods from any release improperly generated during its loading.
-
-        :param release: the target release
-        :return: the release with the stripped disabled mods
-        """
-        # todo: create a new instance for release object and edit it instead, because mutating args is bad mkay?
-        mod_name: str
-        version: ModVersionInfo
-        gtnh_modpack: GTNHModpackManager = await self._get_modpack_manager()
-        github_mods: Dict[str, ModVersionInfo] = release.github_mods
-        external_mods: Dict[str, ModVersionInfo] = release.external_mods
-        valid_side: Set[Side] = {Side.NONE}
-        github_mods_to_delete: List[str] = []
-        external_mods_to_delete: List[str] = []
-        mod_data: Optional[Tuple[GTNHModInfo, GTNHVersion]]
-
-        for mod_name, version in github_mods.items():
-            mod_data = gtnh_modpack.assets.get_mod_and_version(
-                mod_name, version, valid_sides=valid_side, source=ModSource.github
-            )
-            if mod_data is not None:
-                logger.warn(
-                    f"{Fore.YELLOW}Release {release.version} had github mod {mod_name}"
-                    f" in its manifest but it is disabled. Stripping it from memory.{Fore.RESET}"
-                )
-                github_mods_to_delete.append(mod_name)
-
-        for mod_name in github_mods_to_delete:
-            del github_mods[mod_name]
-
-        for mod_name, version in external_mods.items():
-            mod_data = gtnh_modpack.assets.get_mod_and_version(
-                mod_name, version, valid_sides=valid_side, source=ModSource.other
-            )
-            if mod_data is not None:
-                logger.warn(
-                    f"{Fore.YELLOW}Release {self.version} had external mod {mod_name}"
-                    f"in its manifest but it is disabled. Stripping it from memory.{Fore.RESET}"
-                )
-                external_mods_to_delete.append(mod_name)
-
-        for mod_name in external_mods_to_delete:
-            del external_mods[mod_name]
-
-        release.github_mods = github_mods
-        release.external_mods = external_mods
-        return release
+            self.modpack_list_frame.modpack_list.set_loaded_version(self.controller.version)
 
     async def add_gtnh_version(self, release_name: str, previous_version: str) -> None:
         """
@@ -958,33 +479,8 @@ class Window(ThemedTk, Tk):
         :param previous_version: the previous modpack version
         :return: None
         """
-
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        release: GTNHRelease = GTNHRelease(
-            version=release_name,
-            config=self.gtnh_config,
-            github_mods={
-                mod_name: ModVersionInfo(
-                    version=info.version, side=info.side if info.side else gtnh.assets.get_mod(mod_name).side
-                )
-                for mod_name, info in self.github_mods.items()
-            },
-            external_mods={
-                mod_name: ModVersionInfo(
-                    version=info.version, side=info.side if info.side else gtnh.assets.get_mod(mod_name).side
-                )
-                for mod_name, info in self.external_mods.items()
-            },
-            last_version=previous_version,
-        )
-
-        self.last_version = previous_version
-
-        is_release_added: bool = gtnh.add_release(release, update=True)
-
-        if is_release_added:
-            await self.load_gtnh_version(release)
-            gtnh.save_modpack()
+        if await self.controller.add_gtnh_version(release_name, previous_version):
+            showinfo("version loaded successfully!", f"modpack version {release_name} loaded successfully!")
             showinfo("release successfully generated", f"modpack version {release_name} successfully generated!")
 
     async def delete_gtnh_version(self, release_name: str) -> None:
@@ -994,12 +490,8 @@ class Window(ThemedTk, Tk):
         :param release_name: name of the release
         :return: None
         """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        if self.version == release_name:
-            releases: List[GTNHRelease] = await self.get_releases()
-            await self.load_gtnh_version(releases[-1], init=True)
-
-        gtnh.delete_release(release_name)
+        await self.controller.delete_gtnh_version(release_name)
+        self.modpack_list_frame.modpack_list.set_loaded_version(self.controller.version)
         showinfo("release successfully deleted", f"modpack version {release_name} successfully deleted!")
 
     def show(self) -> None:
@@ -1049,25 +541,6 @@ class Window(ThemedTk, Tk):
             self.exclusion_frame_client.show()
             self.exclusion_frame_server.show()
 
-    async def get_external_modlist(self) -> List[str]:
-        """
-        Method to get all the external mods from the assets.
-
-        :return: a list of string with all the external mods availiable
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        return [mod.name for mod in gtnh.assets.mods if mod.source != ModSource.github]
-
-    async def get_modpack_versions(self) -> List[str]:
-        """
-        Method used to gather all the version of the GT-New-Horizons-Modpack repo.
-
-        :return: a list of all the versions availiable.
-        """
-        gtnh: GTNHModpackManager = await self._get_modpack_manager()
-        modpack_config: GTNHConfig = gtnh.assets.config
-        return [version.version_tag for version in modpack_config.versions]
-
     async def run(self) -> None:
         """
         async entrypoint to trigger the mainloop
@@ -1086,24 +559,31 @@ class Window(ThemedTk, Tk):
         if not self.init:
             self.init = True
             # load last gtnh version if there is any:
-            releases: List[GTNHRelease] = await self.get_releases()
+            releases: List[GTNHRelease] = await self.controller.get_releases()
             if len(releases) > 0:
                 await self.load_gtnh_version(releases[-1], init=True)
 
             data_github_mods: Dict[str, Any] = {
-                "github_mod_list": await self.get_repos(),
-                "modpack_version_frame": {"combobox": await self.get_modpack_versions(), "stringvar": self.gtnh_config},
+                "github_mod_list": await self.controller.get_repos(),
+                "modpack_version_frame": {
+                    "combobox": await self.controller.get_modpack_versions(),
+                    "stringvar": self.controller.gtnh_config,
+                },
             }
 
             self.github_panel.populate_data(data_github_mods)
 
-            data_external_mods: Dict[str, Any] = {"external_mod_list": await self.get_external_modlist()}
+            data_external_mods: Dict[str, Any] = {"external_mod_list": await self.controller.get_external_modlist()}
 
             self.external_mod_frame.populate_data(data_external_mods)
 
-            self.modpack_list_frame.populate_data(await self.get_releases())
-            self.exclusion_frame_server.populate_data({"exclusions": await self.get_modpack_exclusions("server")})
-            self.exclusion_frame_client.populate_data({"exclusions": await self.get_modpack_exclusions("client")})
+            self.modpack_list_frame.populate_data(await self.controller.get_releases())
+            self.exclusion_frame_server.populate_data(
+                {"exclusions": await self.controller.get_modpack_exclusions("server")}
+            )
+            self.exclusion_frame_client.populate_data(
+                {"exclusions": await self.controller.get_modpack_exclusions("client")}
+            )
         while self._run:
             self.update()
             self.update_idletasks()
@@ -1115,7 +595,7 @@ class Window(ThemedTk, Tk):
 
         :return: None
         """
-        data_external_mods: Dict[str, Any] = {"external_mod_list": await self.get_external_modlist()}
+        data_external_mods: Dict[str, Any] = {"external_mod_list": await self.controller.get_external_modlist()}
 
         self.external_mod_frame.populate_data(data_external_mods)
 
@@ -1125,9 +605,8 @@ class Window(ThemedTk, Tk):
 
         :return: None
         """
-        if self._client is not None:
-            await self._client.aclose()
-            self._run = False
+        await self.controller.close()
+        self._run = False
         self.destroy()
